wrap in singleton
